+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New: dark theme via `bulma-prefers-dark`
 - Break: use `crossbeam_channel` for message passing so that other services could listen to readings too
+- Break: use service IDs as sensor prefixes
 
 ## `0.8.0`